edits on the main script
because b2 is slow as hell and even if in the settings there is "keep
only last version", it still display multiple versions.

so as a workaroud i load the snapshot with _new in the name, remove the
old one and rename the new one.

also before the snapshot i clean any old snapshot that have _new in the
name. I there are any it means that the backup was interrupted before it
could finish.